safe responsewriter usage in TryExec (#1490)
inside of TryExec we were writing directly to the response writer inside of a
goroutine, but TryExec can timeout and then get called again to a different
runner or even have the front end writing headers while TryExec is writing
headers.

one way to make this safe is to make a new response writer for TryExec to
write the response into, and only after the goroutine handling the response
has returned, from the TryExec goroutine we can copy the response back up as
the caller will not call TryExec again until it has returned (this is
seemingly part of the placer contract). unfortunately, we're already buffering
the response writer in the front end, too - it's possible we can get rid of
that but it may need further testing.

this adds an optimization when copying the request body from the LB to a
runner, since we're using request.GetBody() and returning a reader we
are familiar with that happens to just wrap a buffer's bytes (which we just
need multiple readers on, but the data doesn't change). anyway, this whole
interaction is unfortunate but kind of necessary due to needing to maneuver
into a protobuf, it seems like a worth it and somewhat ok abstraction wise
optimization.

additionally, this gets rid of passing the client response headers down into
the agent for detached functions. we don't need these since detached functions
are not responding with the functions response to the client, only a 202, this
was leading to races around writing the headers in retries too, but this is
just for posterity/correctness now.

updated the makefile/system test script so that I could run these faster to
repro, pretty handy, should add to other stuff too...

closes #1484