replace the timer instead of resetting
since the Reset contract for the timer is messed up, if we wrap the timer and
implement reset by just making a new timer we can get the behavior we want
without having to dance around in the caller making sure we stop timers in any
of our loops before resetting them, since Reset is supposed to behave this
way anyway. and with this, so long as we defer stop it will stop the last
assigned timer. so I think this checks out