removes layer violating ContainerOptions()
this adds all of the options that are required by the most important 3rd party
dependency of all to allow the drivers.ContainerTask to configure any
container create options independently of relying on the underlying driver.
aside from code quality issues and layer violation issues, this allows better
driver portability and maintainability, as well as allows easier switching of
the actually libraries that underlie a given driver (this hack drove past 4
layers of abstraction, and while I can appreciate the dissent on 4 layers of
abstraction existing, it's a pretty bad violation because it creates coupling
issues that drive past all 4 layers).

every use that i've seen is capable of being done via the
drivers.ContainerTask in order to configure a container, simply by decorating
a container task with a new one.