remove unused extensions cruft (#1481)
* changes v1 to v2 for adding an endpoint
* removed the handler funcs for adding handlers onto eg /apps/:app_id/x, we
don't have them for funcs or triggers, and they honestly seem useless as it's
easy to build it with the ability to add a handler and to access the fnext
datastore which we offer, as they were they are really expensive since they
yank the app out of the db even if the operation may not even need it in that
handler. so instead of adding for the rest, remove all of these. none of our
example extensions, which aren't working at the moment, use these either
(that's why I'm here anyway).
* removes dead code helpers and references to app name url param which is no
longer a thing. these were just hanging around bugging me when I ran into
them, so killing them..