Backout #1509; needs more work. (#1511)
The commit "Support specifying image digest on functions (#1509)"
attempted to fix a problem in the runner that occurs when functions that
include both a digest and a tag are invoked.  In certain cases, the
invocation will fail.  This change fixes one aspect of the problem: the
image string parsing.  However, it overlooks other areas where support
for image digests are needed.  This includes the image cache and the
inspection logic.

To further confuse matters, this change makes it harder to observe and
reproduce the original problem.  This makes it difficult to fully
understand both all of the causes and all of the places that require
fixing up.  Remove this for now with the understanding that it will
probably be required as part of a larger fix to simultaneously handle
tags and digests correctly.