Bugfix: use the most recent call for ephemeral data
Things like pull tokens (which can be attached to an incoming call) may
be effectively ephemeral - they have a shorter lifetime than the
slotQueue.

Rather than having hotLauncher use the *original* call that spawned it
for this metadata, we update the slotQueue to hold an atomic reference
to the *most recent* call.

Any tokens on new calls will be used by preference, therefore - meaning
that we won't see a spurious expired token used for image pulls, etc.

Cost here is a single atomic Store/Load, which is probably as good as
it can get. The call extensions should be immutable for this process.