Enforce more rigorous handling of input data and pipes.
Catches and generates function errors for two new cases.  The first
occurs due to a function/FDK error.  If the function closes the read end
of the pipe that the hostagent uses to write data before the hostagent
has finished writing the data, generate an error.  This ensures that
any premature close of the input stream is detected and handled by the
hostagent.

Second, catch any cases where the function attempts to respond before
reading all of the input data from the stream.  This is safe because we
already enforce a maximum upper bound on the request body, so a function
or fdk will not have to read for an unbounded amount of time to consume
the outstanding data.  Since the container contract enforces HTTP-like
semantics, and HTTP expects the server side to wait for the response
body to arrive before responding, this is not unreasonable.  If the
function or fdk attempts to write before the end of the input stream has
been processed by the hostagent, return a different function error
indicating that a premature write has been detected.