fn: slot error retention is suboptimal
Current fn agent (aka runner) tries to communicate
all container start errors to clients. In order to
achieve this, the errors are retained and could be
delivered to clients that did not spawn that container.
This is OK as it tries to be transparent to callers
instead of suppressing or hiding errors. However,
these errors are retained without a time bound which
is not ideal. An old request could trigger an error
and this error can be sent to a client much later time.

With this change, the error retention semantics change.
If an errors occurs during container start and the client
which triggered the container is no longer present, we
log and discard the result.

Broadly, any error that occurs when a request is not 1-1
bound to a container is logged and discarded.

For testing these scenarios, a debug option is added
to agent to allow passing request id that triggered
the spawn of the container as an environment variable.