fn: simplify CPU/Memory acquisition to avoid race condition
Before this change, agent used GetResourceToken() calls
for both non-blocking and blocking mode. However, for
nested calls such as in agent checkLaunch(), a token might
already be queued via go-routine in getResourceTokenNBChan().
If the consuming code in checkLaunch() runs faster, than
it could place another GetResourceToken() call while the
active token is not yet closed. This can result in momentarily
double cpu/mem allocation resulting in 503 or excess wait.

With this PR, resource allocation blocking and non-blocking
interface is changed and ResourceToken is directly returned
without a channel.