Let runner clients specify a timeout at creation.
Add a new function, NewgRPCRunnerWithTimeout, that allows the caller to
define the grpc connection timeout.  Modify NewgRPCRunner to be
implemented in terms of NewgRPCRunnerWithTimeout, passing its existing
default timeout to the new function.

Addition of this function was necessary because the timeout here is
rather aggressive.  A test tool that leverages the existing
NewgRPCRunner was moved into an environment where its test target was in
a different region from the origin, which required a higher timeout to
cope with the test endpoint being physically farther away.